Review: convert a hook without hook usage to function
