+++ frontend/src/metabase/containers/AddToDashSelectDashModal/utils.ts
@@ -1,0 +1,30 @@
+import type { CollectionId, Dashboard } from "metabase-types/api";
+import { ROOT_COLLECTION } from "metabase/entities/collections";
+import { coerceCollectionId } from "metabase/collections/utils";
+
+interface GetInitialOpenCollectionIdProps {
+  isQuestionInPersonalCollection: boolean;
+  mostRecentlyViewedDashboard: Dashboard | undefined;
+}
+
+export const getInitialOpenCollectionId = ({
+  isQuestionInPersonalCollection,
+  mostRecentlyViewedDashboard,
+}: GetInitialOpenCollectionIdProps): undefined | CollectionId => {
+  if (!mostRecentlyViewedDashboard) {
+    return undefined;
+  }
+
+  if (
+    isQuestionInPersonalCollection &&
+    isInPublicCollection(mostRecentlyViewedDashboard)
+  ) {
+    return ROOT_COLLECTION.id;
+  }
+
+  return coerceCollectionId(mostRecentlyViewedDashboard.collection_id);
+};
+
+function isInPublicCollection(dashboard: Dashboard | undefined) {
+  return !dashboard?.collection?.is_personal;
+}
